Use usual file structure for tiller resource
